Break out individual history event parsing into files by opcode/type
